Move ExprExt to ast module
