+++ crates/djls-project/src/extraction.rs
@@ -4,13 +4,11 @@
 //! text and answer star-import recursion through `SettingsSourceResolver`. This
 //! module does not read files, resolve search paths, or depend on Salsa.
 
-pub(crate) mod ext;
 mod extractor;
 mod paths;
 pub(crate) mod registry;
 mod settings;
 
-pub use ext::ExprExt;
 pub use extractor::extract_settings;
 pub use registry::RegistrationInfo;
 pub use registry::RegistrationKind;
