+++ crates/djls-project/src/specs/blocks.rs
@@ -9,7 +9,7 @@ use ruff_python_ast::ExprCall;
 use ruff_python_ast::ExprName;
 use ruff_python_ast::StmtFunctionDef;
 
-use crate::extraction::ext::ExprExt;
+use crate::ast::ExprExt;
 use crate::specs::types::BlockSpec;
 
 /// Extract a block spec from a tag's compile function.
