+++ crates/djls-project/src/specs/blocks/dynamic_end.rs
@@ -12,7 +12,7 @@ use ruff_python_ast::StmtReturn;
 use ruff_python_ast::statement_visitor::StatementVisitor;
 use ruff_python_ast::statement_visitor::walk_stmt;
 
-use crate::extraction::ext::ExprExt;
+use crate::ast::ExprExt;
 use crate::specs::blocks::is_parser_receiver;
 use crate::specs::types::BlockSpec;
