+++ crates/djls-project/src/specs/blocks/parse_calls.rs
@@ -8,7 +8,7 @@ use ruff_python_ast::statement_visitor::StatementVisitor;
 use ruff_python_ast::statement_visitor::walk_body;
 use ruff_python_ast::statement_visitor::walk_stmt;
 
-use crate::extraction::ext::ExprExt;
+use crate::ast::ExprExt;
 use crate::specs::blocks::extract_string_sequence;
 use crate::specs::blocks::is_parser_receiver;
 use crate::specs::blocks::is_token_contents_expr;
